+++ DONNEES/Base_Donnee.h
@@ -4,4 +4,4 @@ int initDataTable(int N_COLS, int* table_donnees);
 
 void loadDataTable(int N_COLS, int* table_donnees, FILE* lastGame);
 
-void loadVariables(int position, FILE* lastGame, int * variable);
+void loadVariables(long position, FILE* lastGame, int * variable);
